feat: create organisation (#2254)
* feat: refactor dashboard UI frame

* refactor: router. use router with context

* refactor: remove unused tailwind classes

* fix: sidebar appearing over main content

* feat: use auth hook

* feat: use auth hook

* refactor: auth service

* fix: stop caching useAuth return values

* fix: login button not enable on request completion on error

* feat: set overlay background colour for dialog component

* feat: update sidebar background colour

* feat: create organisation
