+++ web/ui/dashboard-react/src/app/index.tsx
@@ -1,10 +1,11 @@
 import { createFileRoute } from '@tanstack/react-router';
+
 import { router } from '@/lib/router';
 import { ensureCanAccessPrivatePages } from '@/lib/auth';
 
 export const Route = createFileRoute('/')({
-	beforeLoad() {
-		ensureCanAccessPrivatePages();
+	beforeLoad({context}) {
+		ensureCanAccessPrivatePages(context.auth?.getTokens().isLoggedIn);
 		router.navigate({ to: '/projects' });
 	},
 	component: Index,
