+++ web/ui/dashboard-react/src/app/projects/_layout.tsx
@@ -1,0 +1,19 @@
+import { ensureCanAccessPrivatePages } from '@/lib/auth';
+import { Outlet, createFileRoute } from '@tanstack/react-router';
+
+import { DashboardLayout } from '@/components/dashboard';
+
+export const Route = createFileRoute('/projects/_layout')({
+	beforeLoad({ context }) {
+		ensureCanAccessPrivatePages(context.auth?.getTokens().isLoggedIn);
+	},
+	component: ProjectsLayout,
+});
+
+function ProjectsLayout() {
+	return (
+		<DashboardLayout>
+			<Outlet />
+		</DashboardLayout>
+	);
+}
