+++ web/ui/dashboard-react/src/lib/router.ts
@@ -1,7 +1,7 @@
 import { createRouter } from '@tanstack/react-router';
 import { routeTree } from '../routes.gen';
 
-export const router = createRouter({ routeTree });
+export const router = createRouter({ routeTree, context: { auth: null } });
 
 declare module '@tanstack/react-router' {
 	interface Register {
