+++ web/ui/dashboard-react/src/main.tsx
@@ -1,9 +1,10 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { RouterProvider } from '@tanstack/react-router';
-import { router } from './lib/router';
 
 import './index.css';
+import { router } from './lib/router';
+import { useAuth } from './hooks/use-auth';
 
 let root = document.getElementById('root');
 if (!root) {
@@ -13,8 +14,13 @@ if (!root) {
 	root = document.getElementById('root') as HTMLElement;
 }
 
+function App() {
+	const auth = useAuth();
+	return <RouterProvider router={router} context={{ auth }} />;
+}
+
 createRoot(root).render(
 	<StrictMode>
-		<RouterProvider router={router} />
+		<App />
 	</StrictMode>,
 );
