+++ web/ui/dashboard-react/tailwind.config.js
@@ -180,32 +180,6 @@ module.exports = {
 				800: '#05603A,',
 				900: '#054F31,',
 			},
-			'new.error': {
-				25: '#FFFBFA',
-				50: '#FEF3F2',
-				100: '#FEE4E2',
-				200: '#FECDCA',
-				300: '#FDA29B',
-				400: '#F97066',
-				500: '#F04438',
-				600: '#D92D20',
-				700: '#B42318',
-				800: '#912018,',
-				900: '#7A271A,',
-			},
-			'new.warning': {
-				25: '#FFFCF5',
-				50: '#FFFAEB',
-				100: '#FEF0C7',
-				200: '#FEDF89',
-				300: '#FEC84B',
-				400: '#FDB022',
-				500: '#F79009',
-				600: '#DC6803',
-				700: '#B54708',
-				800: '#93370D,',
-				900: '#7A2E0E,',
-			},
 			'new.gray': {
 				25: '#FCFCFD',
 				50: '#F9FAFB',
@@ -220,7 +194,6 @@ module.exports = {
 				900: '#101828',
 			},
 			neutral: generateColorScale('gray'),
-			error: generateColorScale('red'),
 			white: {
 				4: 'rgba(var(--color-white), 0.04)',
 				8: 'rgba(var(--color-white), 0.08)',
